Update TeamLDR_CS470_Project4.py
Nearly completed, however, we still have some issue with checkMultiJump. It is not accounting for all available potential jumps. One pro, the ones it does highlight are 100% valid. Forgot to account for pieces not able to go back into starting point and cannot leave when it enters goal area already...Need to add a timer/countdown eventually and put it in a label (maybe).